move code from game to separate chat module
